Reversed order of postprocessing of the log files. The error.log is now handled first, since it is likely to be smaller. The rsync.log is handled immideately after.
Compression of the log files is now prioritized by prepending a nice command to the compression. We can always lower the task priority, only root can increase it. See man nice(1) for more info about task nice levels.

If you're trying to increase the task priority but are not running as root, this will be noted in the admin message.

MLbeNice defaults to 20 (lowest possible priority) to be nice to other processes on the system. Postprocessing the logs is not considered vital, since the backup is already done at that time.


If the admin notification message is not empty it will also be moved over to the backup destination (on localhost only at the moment). The same message should be emailed to the admin in a future version.