Added some more debug output for the temporary files, so one can trace them during testing.
Hopefully fixed a BAD bug, that caused rsync to background on longer backups and do all kinds of weeird stuff in the premature post production of the log files etc. Just a dumb typo, should be solved now.

Removed -p from target directory creation. Should post an error causing the script to exit if an external Target Volume is not present during backup. Should be handled more gracefully though. That means we should check this properly first.
This is a quick and very dirty hack, it will very likely NOT give the intended result.

Changed tempfile creation of the AdminEMailMessage from tempfile to mktemp. An oversight of mine when changing the other tempfiles. Fixed now.

Fixed a bug where the destination rsync parameter was passed as empty. Didn't yet affect us since we're still in localhost only mode. Remote backups are not yet really supported by the script.