Nice refactoring of the Backup rotation code. Now will NOT hang in an infinite loop anymore if a backup cannot be deleted. Takes care of any changes in the backup configuration. Will remove more than the oldest backup if the backup count has been reduced in the config file.
Debug Output is more informative and correct now. Should really help to track down problems if you have some.

Properly builds a list of old surplus backups and correctly reports them in debug. Gices more info in the Adminmail. (still some ideas to improve on this).

Still needs to check if we actually succeeded in removing the old surplus backups.