This BUILD is OK.
Fixed a bug, that caused the red label and comment to fail when the backup itself failed as well. Now works correctly. 

Added even more debug lines to help aid debugging problems.

Mail Messages are correctly sent now no matter what the backup status is. Archiving of logs and messages to the backup works as well.