Upped the MLversion to 0.6 since we're in the 6x revisions.
Out Commented all the growl stuff for the moment, so we can better concentrate on the core functionality of the backup. Growl support will be re-added later on.
Then it is likely to support running on a server, remote growl, as well as local growl if run on a client machine.

introduced new Variable for backup postprocessing:
MLfatalError = 1
indicated that an error returned by rsync is to be considered fatal for the backup and hence it should not be used.

MLfatalError = 2
Indicates that a non serious error has occured, aka warning. This is the case for vanished source files, for example. An error that frequently occurs when the backup takes place during user interaction with the source. Usually not hazardous to the backup process at all, hence considered non fatal, but a warning.

MLfatalError = 0
Indicates that no fatal error has occured at all, and the backup is considered sound.


Next Step is to move the postprocessing out of the case block to only decide what is to be done afterwards.