Major moving of logic blocks. Now if an error occured in rsync we decide if it was fatal or not. With fatal errors we act like we did up until now. non Fatal errors will be treated as if the backup went ok.
We prepared for handling Fatal errors like they should be, minor errors like a warning and no errors like they should be. :-)

This still has to be tested a LOT until we can release this again as mostly useable. I just need to check over the logic again, if there aren't any misbehaviours.