Fixed a bug that lead to error messages like:
head: illegal option -- -
Usage: head [-n lines] [file ...]

when there were no backups to delete because the current number of backups was less than MLbackupCount
so a negative number was passed to head. Reordered the logic so this evaluation only takes place when there definitely are old backups to delete. This gets rid of the error message.