Added extended text encoding options to help with diacritic characters in filenames. Mac users usually use a multitude of these. Using the --iconv option should help reduce the transmission of already-up-to-date files and hence speed up the backup process, especially when creating remote backups using ssh transport.
From the rsync(1) 3 man page:
--iconv=UTF-8,UTF8-MAC

An example of the latter can occur with HFS+ on Mac OS X: if you copy a directory with a file that has a UTF-8 character sequence in it, say a 2-byte umlaut-u (\0303\0274), the file will get that character stored by the filesystem using 3 bytes (\0165\0314\0210), and rsync will not know that these differing filenames are the same file (it will, in fact, remove a prior copy of the file if --delete is enabled, and then recreate it).

You can avoid a charset problem by passing an appropriate --iconv option to rsync that tells it what character-set the source files are, and what character-set the destination files get stored in. For instance, the above Mac OS X problem would be dealt with by using --iconv=UTF-8,UTF8-MAC (UTF8-MAC is a pseudo-charset recognized by Mac OS X iconv in which all characters are decomposed).